added try and except for completeness profile in PNe_fitting.py
